smart disclaimer
