+++ components/bill/Summary.tsx
@@ -5,6 +5,7 @@ import Image from "react-bootstrap/Image"
 import styled from "styled-components"
 import { useMediaQuery } from "usehooks-ts"
 import { Button, Col, Container, Modal, Row } from "../bootstrap"
+import { SmartDisclaimer } from "./SmartDisclaimer"
 import { TestimonyCounts } from "./TestimonyCounts"
 import { BillProps } from "./types"
 
@@ -43,10 +44,6 @@ const FormattedBillDetails = styled(Col)`
   white-space: pre-wrap;
 `
 
-const SmartTagDesc = styled(Row)`
-  font-size: 12px;
-`
-
 export const Summary = ({
   bill,
   className
@@ -56,7 +53,6 @@ export const Summary = ({
   const handleHideBillDetails = () => setShowBillDetails(false)
   const billText = bill?.content?.DocumentText
 
-  const isMobile = useMediaQuery("(max-width: 768px)")
   const { t } = useTranslation("common")
 
   console.log("Bill: ", bill)
@@ -107,35 +103,7 @@ export const Summary = ({
         </Col>
       </Row>
       <hr className={`m-0 border-bottom border-2`} />
-      {isMobile ? (
-        <div className={`mx-2`}>
-          <div className={`d-flex justify-content-center`}>
-            <Image
-              src="/smart-tag.png"
-              alt={t("bill.smart-tag")}
-              // className={`d-flex justify-self-center`}
-              // height="53"
-              // width="38"
-            />
-          </div>
-          <Row className={`fs-5 fw-bold`}>{t("bill.smart_summary")}</Row>
-          <SmartTagDesc>{t("bill.smart_explanation")}</SmartTagDesc>
-        </div>
-      ) : (
-        <Row className={`d-flex my-3`} xs="auto">
-          <Col>
-            <Image
-              src="/smart-tag.png"
-              alt={t("bill.smart-tag")}
-              className={``}
-            />
-          </Col>
-          <Col className={`mt-1`} xs="10">
-            <Row className={`fs-5 fw-bold`}>{t("bill.smart_summary")}</Row>
-            <SmartTagDesc>{t("bill.smart_explanation")}</SmartTagDesc>
-          </Col>
-        </Row>
-      )}
+      <SmartDisclaimer />
     </SummaryContainer>
   )
 }
